+++ packages/router/__tests__/router.spec.ts
@@ -4,15 +4,15 @@ import {
   createMemoryHistory,
   createWebHistory,
   createWebHashHistory,
+  loadRouteLocation,
+  RouteLocationRaw,
 } from '../src'
 import { NavigationFailureType } from '../src/errors'
 import { createDom, components, tick, nextNavigation } from './utils'
-import {
-  RouteRecordRaw,
-  RouteLocationRaw,
-  START_LOCATION_NORMALIZED,
-} from '../src/types'
-import { mockWarn } from 'jest-mock-warn'
+import { RouteRecordRaw } from '../src/types'
+import { START_LOCATION_NORMALIZED } from '../src/location'
+import { vi, describe, expect, it, beforeAll } from 'vitest'
+import { mockWarn } from './vitest-mock-warn'
 
 declare var __DEV__: boolean
 
@@ -110,7 +110,7 @@ describe('Router', () => {
 
   it('calls history.push with router.push', async () => {
     const { router, history } = await newRouter()
-    jest.spyOn(history, 'push')
+    vi.spyOn(history, 'push')
     await router.push('/foo')
     expect(history.push).toHaveBeenCalledTimes(1)
     expect(history.push).toHaveBeenCalledWith('/foo', undefined)
@@ -119,7 +119,7 @@ describe('Router', () => {
   it('calls history.replace with router.replace', async () => {
     const history = createMemoryHistory()
     const { router } = await newRouter({ history })
-    jest.spyOn(history, 'replace')
+    vi.spyOn(history, 'replace')
     await router.replace('/foo')
     expect(history.replace).toHaveBeenCalledTimes(1)
     expect(history.replace).toHaveBeenCalledWith('/foo', expect.anything())
@@ -128,7 +128,7 @@ describe('Router', () => {
   it('parses query and hash with router.replace', async () => {
     const history = createMemoryHistory()
     const { router } = await newRouter({ history })
-    jest.spyOn(history, 'replace')
+    vi.spyOn(history, 'replace')
     await router.replace('/foo?q=2#a')
     expect(history.replace).toHaveBeenCalledTimes(1)
     expect(history.replace).toHaveBeenCalledWith(
@@ -142,8 +142,8 @@ describe('Router', () => {
     const { router } = await newRouter({ history })
     // move somewhere else
     await router.push('/search')
-    jest.spyOn(history, 'replace')
-    jest.spyOn(history, 'push')
+    vi.spyOn(history, 'replace')
+    vi.spyOn(history, 'push')
     await router.replace('/home-before')
     expect(history.push).toHaveBeenCalledTimes(0)
     expect(history.replace).toHaveBeenCalledTimes(1)
@@ -160,8 +160,8 @@ describe('Router', () => {
       }
       return // no warn
     })
-    jest.spyOn(history, 'replace')
-    jest.spyOn(history, 'push')
+    vi.spyOn(history, 'replace')
+    vi.spyOn(history, 'push')
     await router.push('/search')
     expect(history.location).toBe('/foo')
     expect(history.push).toHaveBeenCalledTimes(0)
@@ -170,7 +170,7 @@ describe('Router', () => {
   })
 
   it('allows to customize parseQuery', async () => {
-    const parseQuery = jest.fn(_ => ({}))
+    const parseQuery = vi.fn(_ => ({}))
     const { router } = await newRouter({ parseQuery })
     const to = router.resolve('/foo?bar=baz')
     expect(parseQuery).toHaveBeenCalledWith('bar=baz')
@@ -178,7 +178,7 @@ describe('Router', () => {
   })
 
   it('allows to customize stringifyQuery', async () => {
-    const stringifyQuery = jest.fn(_ => '')
+    const stringifyQuery = vi.fn(_ => '')
     const { router } = await newRouter({ stringifyQuery })
     const to = router.resolve({ query: { foo: 'bar' } })
     expect(stringifyQuery).toHaveBeenCalledWith({ foo: 'bar' })
@@ -187,7 +187,7 @@ describe('Router', () => {
   })
 
   it('creates an empty query with no query', async () => {
-    const stringifyQuery = jest.fn(_ => '')
+    const stringifyQuery = vi.fn(_ => '')
     const { router } = await newRouter({ stringifyQuery })
     const to = router.resolve({ hash: '#a' })
     expect(stringifyQuery).not.toHaveBeenCalled()
@@ -255,7 +255,7 @@ describe('Router', () => {
 
   it('can pass replace option to push', async () => {
     const { router, history } = await newRouter()
-    jest.spyOn(history, 'replace')
+    vi.spyOn(history, 'replace')
     await router.push({ path: '/foo', replace: true })
     expect(history.replace).toHaveBeenCalledTimes(1)
     expect(history.replace).toHaveBeenCalledWith('/foo', expect.anything())
@@ -263,7 +263,7 @@ describe('Router', () => {
 
   it('can replaces current location with a string location', async () => {
     const { router, history } = await newRouter()
-    jest.spyOn(history, 'replace')
+    vi.spyOn(history, 'replace')
     await router.replace('/foo')
     expect(history.replace).toHaveBeenCalledTimes(1)
     expect(history.replace).toHaveBeenCalledWith('/foo', expect.anything())
@@ -271,7 +271,7 @@ describe('Router', () => {
 
   it('can replaces current location with an object location', async () => {
     const { router, history } = await newRouter()
-    jest.spyOn(history, 'replace')
+    vi.spyOn(history, 'replace')
     await router.replace({ path: '/foo' })
     expect(history.replace).toHaveBeenCalledTimes(1)
     expect(history.replace).toHaveBeenCalledWith('/foo', expect.anything())
@@ -279,7 +279,7 @@ describe('Router', () => {
 
   it('navigates if the location does not exist', async () => {
     const { router } = await newRouter({ routes: [routes[0]] })
-    const spy = jest.fn((to, from, next) => next())
+    const spy = vi.fn((to, from, next) => next())
     router.beforeEach(spy)
     await router.push('/idontexist')
     expect(spy).toHaveBeenCalledTimes(1)
@@ -468,7 +468,9 @@ describe('Router', () => {
     expect(
       router.resolve(
         { params: { p: 1 } },
-        router.resolve({ name: 'Param', params: { p: 2 } })
+        await loadRouteLocation(
+          router.resolve({ name: 'Param', params: { p: 2 } })
+        )
       )
     ).toMatchObject({
       name: 'Param',
@@ -502,7 +504,7 @@ describe('Router', () => {
   describe('alias', () => {
     it('does not navigate to alias if already on original record', async () => {
       const { router } = await newRouter()
-      const spy = jest.fn((to, from, next) => next())
+      const spy = vi.fn((to, from, next) => next())
       await router.push('/basic')
       router.beforeEach(spy)
       await router.push('/basic-alias')
@@ -511,7 +513,7 @@ describe('Router', () => {
 
     it('does not navigate to alias with children if already on original record', async () => {
       const { router } = await newRouter()
-      const spy = jest.fn((to, from, next) => next())
+      const spy = vi.fn((to, from, next) => next())
       await router.push('/aliases')
       router.beforeEach(spy)
       await router.push('/aliases1')
@@ -522,7 +524,7 @@ describe('Router', () => {
 
     it('does not navigate to child alias if already on original record', async () => {
       const { router } = await newRouter()
-      const spy = jest.fn((to, from, next) => next())
+      const spy = vi.fn((to, from, next) => next())
       await router.push('/aliases/one')
       router.beforeEach(spy)
       await router.push('/aliases1/one')
@@ -697,7 +699,7 @@ describe('Router', () => {
     it('only triggers guards once with a redirect option', async () => {
       const history = createMemoryHistory()
       const router = createRouter({ history, routes })
-      const spy = jest.fn((to, from, next) => next())
+      const spy = vi.fn((to, from, next) => next())
       router.beforeEach(spy)
       await router.push('/to-foo')
       expect(spy).toHaveBeenCalledTimes(1)