+++ packages/router/src/matcher/types.ts
@@ -1,9 +1,11 @@
+import type {
+  NavigationGuard,
+  NavigationGuardNextCallback,
+  _RouteRecordProps,
+} from '../typed-routes'
 import {
   RouteRecordMultipleViews,
-  NavigationGuard,
   _RouteRecordBase,
-  _RouteRecordProps,
-  NavigationGuardNextCallback,
   RouteRecordRaw,
 } from '../types'
 import { ComponentPublicInstance } from 'vue'