+++ packages/router/src/typed-routes/index.ts
@@ -1,0 +1,5 @@
+export type * from './params'
+export type * from './route-map'
+export type * from './route-location'
+export type * from './route-records'
+export type * from './navigation-guards'