+++ packages/router/src/types/typeGuards.ts
@@ -1,9 +1,9 @@
-import { RouteLocationRaw, RouteRecordName } from './index'
+import type { RouteLocationRaw, RouteRecordNameGeneric } from '../typed-routes'
 
 export function isRouteLocation(route: any): route is RouteLocationRaw {
   return typeof route === 'string' || (route && typeof route === 'object')
 }
 
-export function isRouteName(name: any): name is RouteRecordName {
+export function isRouteName(name: any): name is RouteRecordNameGeneric {
   return typeof name === 'string' || typeof name === 'symbol'
 }