+++ packages/router/src/useApi.ts
@@ -1,7 +1,8 @@
 import { inject } from 'vue'
 import { routerKey, routeLocationKey } from './injectionSymbols'
 import { Router } from './router'
-import { RouteLocationNormalizedLoaded } from './types'
+import { RouteMap } from './typed-routes/route-map'
+import { RouteLocationNormalizedLoaded } from './typed-routes'
 
 /**
  * Returns the router instance. Equivalent to using `$router` inside
@@ -15,6 +16,8 @@ export function useRouter(): Router {
  * Returns the current route location. Equivalent to using `$route` inside
  * templates.
  */
-export function useRoute(): RouteLocationNormalizedLoaded {
+export function useRoute<Name extends keyof RouteMap = keyof RouteMap>(
+  _name?: Name
+): RouteLocationNormalizedLoaded<Name> {
   return inject(routeLocationKey)!
 }