+++ packages/router/test-dts/meta.test-d.ts
@@ -1,5 +1,6 @@
-import { createRouter, createWebHistory, expectType } from './index'
-import { createApp, defineComponent } from 'vue'
+import { createRouter, createWebHistory } from './index'
+import { defineComponent } from 'vue'
+import { describe, it, expectTypeOf } from 'vitest'
 
 const component = defineComponent({})
 
@@ -10,34 +11,41 @@ declare module './index' {
   }
 }
 
-const router = createRouter({
-  history: createWebHistory(),
-  routes: [
-    {
-      path: '/',
-      component,
-      meta: {
-        requiresAuth: true,
-        lol: true,
-        nested: {
-          foo: 'bar',
+describe('RouteMeta', () => {
+  it('route creation', () => {
+    const router = createRouter({
+      history: createWebHistory(),
+      routes: [
+        {
+          path: '/',
+          component,
+          meta: {
+            requiresAuth: true,
+            lol: true,
+            nested: {
+              foo: 'bar',
+            },
+          },
         },
-      },
-    },
-    {
-      path: '/foo',
-      component,
-      // @ts-expect-error
-      meta: {},
-    },
-  ],
-})
+        {
+          path: '/foo',
+          component,
+          // @ts-expect-error
+          meta: {},
+        },
+      ],
+    })
+  })
 
-router.beforeEach(to => {
-  expectType<{ requiresAuth?: Boolean; nested: { foo: string } }>(to.meta)
-  expectType<unknown>(to.meta.lol)
-  if (to.meta.nested.foo == 'foo' || to.meta.lol) return false
+  it('route location in guards', () => {
+    const router = createRouter({
+      history: createWebHistory(),
+      routes: [],
+    })
+    router.beforeEach(to => {
+      expectTypeOf<{ requiresAuth?: Boolean; nested: { foo: string } }>(to.meta)
+      expectTypeOf<unknown>(to.meta.lol)
+      if (to.meta.nested.foo == 'foo' || to.meta.lol) return false
+    })
+  })
 })
-
-const app = createApp({})
-app.use(router)